refactor: Rename AttachementMixin to AttachmentMixin
