+++ src/mixins/AttachmentMixin.js
